% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{optimalr}
\alias{optimalr}
\alias{optimalr-package}
\title{Optimization Snippets for R}
\description{
Optimization Snippets for R
}
\section{Package options}{


optimalr uses the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{digits}: number of digits to print
  \item \code{expressions}: a limit on the number of nested exp to eva
  \item \code{max.print}: max into to print
}
}

